[MNG-8411] The mvnenc tool goals (sans init) should work in batch mode
This makes the 3 goals sans init work in batch mode.

---

https://issues.apache.org/jira/browse/MNG-8411
